backend_api/requirements.txt
# Backend API - AI Photobooth Kota Madiun
# Python FastAPI + Supabase

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0

# Supabase
supabase==2.7.0

# ML - Background Removal
Pillow==10.4.0
rembg==2.0.57
numpy==1.26.4
opencv-python-headless==4.10.0.84

# ML - Face Expression (YOLO)
ultralytics==8.2.0
torch==2.4.0
torchvision==0.19.0

# ML - QR Code
qrcode[pil]==7.4.2

# HTTP
requests==2.32.0
httpx==0.27.0

# Utils
aiofiles==24.1.0
